requirements.txt
python-telegram-bot[ext]==21.*
httpx==0.27.*
aiosqlite==0.20.*
python-dotenv==1.0.*
